Move compute requirements under single package
