+++ skyplane/chunk.py
@@ -2,6 +2,7 @@ import socket
 from dataclasses import asdict, dataclass
 from enum import Enum, auto
 from functools import total_ordering
+
 from typing import Dict, Optional
