+++ skyplane/cli/cli.py
@@ -4,12 +4,12 @@ import time
 import traceback
 from importlib.resources import path
 from shlex import split
-from typing import Optional
 
 import typer
 from rich import print as rprint
 from rich.progress import Progress, SpinnerColumn, TextColumn
 from rich.prompt import IntPrompt
+from typing import Optional
 
 import skyplane.cli
 import skyplane.cli
@@ -31,7 +31,6 @@ from skyplane.cli.cli_impl.cp_replicate import (
     launch_replication_job,
 )
 from skyplane.cli.cli_impl.cp_replicate_fallback import (
-    get_usage_gbits,
     replicate_onprem_cp_cmd,
     replicate_onprem_sync_cmd,
     replicate_small_cp_cmd,
@@ -40,12 +39,12 @@ from skyplane.cli.cli_impl.cp_replicate_fallback import (
 from skyplane.cli.cli_impl.init import load_aws_config, load_azure_config, load_gcp_config
 from skyplane.cli.common import console, print_header, print_stats_completed, query_instances
 from skyplane.cli.usage.client import UsageClient, UsageStatsStatus
-from skyplane.compute.aws.aws_auth import AWSAuthentication
-from skyplane.compute.aws.aws_cloud_provider import AWSCloudProvider
-from skyplane.compute.azure.azure_auth import AzureAuthentication
-from skyplane.compute.azure.azure_cloud_provider import AzureCloudProvider
-from skyplane.compute.gcp.gcp_auth import GCPAuthentication
-from skyplane.compute.gcp.gcp_cloud_provider import GCPCloudProvider
+from skyplane.compute import AWSAuthentication
+from skyplane.compute import AWSCloudProvider
+from skyplane.compute import AzureAuthentication
+from skyplane.compute import AzureCloudProvider
+from skyplane.compute import GCPAuthentication
+from skyplane.compute import GCPCloudProvider
 from skyplane.config import SkyplaneConfig
 from skyplane.config_paths import config_path, cloud_config
 from skyplane.obj_store.object_store_interface import ObjectStoreInterface
