+++ skyplane/cli/cli_impl/cp_replicate_fallback.py
@@ -1,4 +1,5 @@
 import os
+
 from typing import Optional
 
 from skyplane.api.impl.path import parse_path
