+++ skyplane/compute/azure/azure_auth.py
@@ -2,6 +2,7 @@ import json
 import logging
 import os
 import subprocess
+
 from typing import Dict, List, Optional
 
 from skyplane.compute.const_cmds import query_which_cloud
