+++ skyplane/compute/gcp/gcp_auth.py
@@ -1,6 +1,7 @@
 import base64
 import os
 from pathlib import Path
+
 from typing import Optional
 
 from skyplane.compute.server import key_root
