+++ skyplane/gateway/chunk_store.py
@@ -4,6 +4,7 @@ from multiprocessing import Manager, Queue
 from os import PathLike
 from pathlib import Path
 from queue import Empty
+
 from typing import Dict, List, Optional
 
 from skyplane.chunk import ChunkRequest, ChunkState
