+++ skyplane/gateway/gateway_daemon.py
@@ -10,6 +10,7 @@ from multiprocessing import Event, Queue
 from os import PathLike
 from pathlib import Path
 from threading import BoundedSemaphore
+
 from typing import Dict
 
 from skyplane.chunk import ChunkState
