+++ skyplane/gateway/gateway_obj_store.py
@@ -4,6 +4,7 @@ import traceback
 from dataclasses import dataclass
 from functools import partial
 from multiprocessing import Event, Manager, Process, Value, Queue
+
 from typing import Dict, Optional
 
 from skyplane.chunk import ChunkRequest
