+++ skyplane/obj_store/object_store_interface.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+
 from typing import Iterator, List, Optional, Tuple
