+++ skyplane/obj_store/s3_interface.py
@@ -2,10 +2,11 @@ import base64
 import hashlib
 import os
 from functools import lru_cache
+
 from typing import Iterator, List, Optional, Tuple
 
 from skyplane import exceptions
-from skyplane.compute.aws.aws_auth import AWSAuthentication
+from skyplane.compute import AWSAuthentication
 from skyplane.exceptions import NoSuchObjectException
 from skyplane.obj_store.object_store_interface import ObjectStoreInterface, ObjectStoreObject
 from skyplane.utils import logger, imports
