+++ skyplane/replicate/solver.py
@@ -2,12 +2,12 @@ import functools
 import shutil
 from collections import namedtuple
 from dataclasses import dataclass
-from typing import Dict, List, Optional, Tuple
 
 import numpy as np
 import pandas as pd
+from typing import Dict, List, Optional, Tuple
 
-from skyplane.compute.cloud_providers import CloudProvider
+from skyplane.compute import CloudProvider
 from skyplane.replicate.replication_plan import ReplicationTopology
 from skyplane.utils import logger
 from skyplane.utils.definitions import GB
